requirements.txt
# Customs Scraper (CMM)
scrapling>=0.4
apscheduler>=3.10
python-dotenv>=1.0

# Policy Monitor (CPM)
feedparser>=6.0.0
requests>=2.31.0
pyyaml>=6.0
akshare>=1.14.0
flask>=3.0.0
openpyxl>=3.1.0
global-macro-data>=2.0.0
websockets>=12.0

# Test
pytest>=8.0
lxml>=5.0
